auth@useApiDocs

+++ nuxtapp/composables/useApiDocs.ts
@@ -1,12 +1,14 @@
 import { URL_DOCS } from "@/config";
 import type { OrNull, IDoc } from "@/types";
+import { useStoreApiAuth } from "../stores/auth";
 
 export const useApiDocs = <TDocData>(tagName: string) => {
   const {
     io: { IOEVENT_DOCS_CHANGE },
   } = useAppConfig();
   const ioEventDocsChange = `${IOEVENT_DOCS_CHANGE}:${tagName}`;
-  const docsUrl = `${URL_DOCS}/${tagName}`;
+  const URL_DOCS__TAGGED = `${URL_DOCS}/${tagName}`;
+  const KEY_NUXT_DATA_DOCS = `docs:${tagName}`;
 
   const crudToggleProcessing = useToggleFlag();
   const crudProcessingError$ = ref<any>(null);
@@ -15,20 +17,38 @@ export const useApiDocs = <TDocData>(tagName: string) => {
     crudToggleProcessing.on();
   };
 
+  const auth = useStoreApiAuth();
+  const {
+    stores: {
+      auth: { authHeaders },
+    },
+  } = useAppConfig();
+  const authHeaders$ = computed(() =>
+    auth.token$ ? authHeaders(auth.token$) : {}
+  );
+
   const {
     error,
     pending,
     refresh,
     data: docs,
-  } = useFetch<IDoc<TDocData>[]>(docsUrl, {
-    key: `docs:${tagName}`,
+  } = useFetch<IDoc<TDocData>[]>(URL_DOCS__TAGGED, {
+    key: KEY_NUXT_DATA_DOCS,
     lazy: true,
     default: () => [],
+    headers: authHeaders$,
+    immediate: false,
   });
-  const reload = async () => await refresh();
+
   const processing$ = computed(
     () => pending.value || crudToggleProcessing.isActive.value
   );
+
+  const reload = async () => {
+    if (auth.token$) await refresh();
+  };
+  watch(() => auth.token$, reload);
+  onMounted(reload);
 
   useIOEvent(ioEventDocsChange, reload);
 
@@ -36,20 +56,22 @@ export const useApiDocs = <TDocData>(tagName: string) => {
   const put = async (doc: IDoc<TDocData>) => {
     let res: OrNull<IDoc<TDocData>> = null;
 
+    if (!auth.token$) return;
     if (processing$.value) return;
 
     crudProcessingStart();
 
     try {
-      res = await $fetch(docsUrl, {
+      res = await $fetch(URL_DOCS__TAGGED, {
+        headers: authHeaders(auth.token$),
         method: "POST",
         body: doc,
       });
     } catch (err) {
       crudProcessingError$.value = err;
-    } finally {
-      crudToggleProcessing.off();
     }
+
+    crudToggleProcessing.off();
 
     return res;
   };
@@ -58,12 +80,14 @@ export const useApiDocs = <TDocData>(tagName: string) => {
   const rm = async (doc: IDoc<TDocData>) => {
     let res: OrNull<IDoc<TDocData>> = null;
 
+    if (!auth.token$) return;
     if (processing$.value) return;
 
     crudProcessingStart();
 
     try {
-      res = await $fetch(docsUrl, {
+      res = await $fetch(URL_DOCS__TAGGED, {
+        headers: authHeaders(auth.token$),
         method: "DELETE",
         body: {
           id: doc?.id,
@@ -71,9 +95,9 @@ export const useApiDocs = <TDocData>(tagName: string) => {
       });
     } catch (err) {
       crudProcessingError$.value = err;
-    } finally {
-      crudToggleProcessing.off();
     }
+
+    crudToggleProcessing.off();
 
     return res;
   };
